# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Login"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registration",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.Registration"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Registration"
                        }
                    }
                }
            }
        },
        "/v1/flat/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create Flat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flat"
                ],
                "summary": "Create Flat",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.CreateFlat"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Flat"
                        }
                    }
                }
            }
        },
        "/v1/flat/update": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update Flat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flat"
                ],
                "summary": "Update Flat",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.UpdateFlat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Flat"
                        }
                    }
                }
            }
        },
        "/v1/house/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create House",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "house"
                ],
                "summary": "Create House",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.CreateHouse"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.House"
                        }
                    }
                }
            }
        },
        "/v1/house/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get House Flats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "house"
                ],
                "summary": "Get House Flats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.HouseFlats"
                        }
                    }
                }
            }
        },
        "/v1/house/{id}/subscribe": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Subscribe for house updates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "house"
                ],
                "summary": "Subscribe for house updates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.CreateFlat": {
            "type": "object",
            "required": [
                "house_id",
                "number",
                "price",
                "rooms_amount"
            ],
            "properties": {
                "house_id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "rooms_amount": {
                    "type": "integer"
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.CreateHouse": {
            "type": "object",
            "required": [
                "address",
                "year"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "developer": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 4
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.Registration": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 4
                },
                "user_type": {
                    "type": "string",
                    "enum": [
                        "client",
                        "moderator"
                    ]
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_request.UpdateFlat": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "created",
                        "approved",
                        "declined",
                        "on moderation"
                    ]
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Flat": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderation_status": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "rooms_amount": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.House": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "developer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_ats": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.HouseFlats": {
            "type": "object",
            "properties": {
                "flats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_romanchechyotkin_avito_test_task_internal_entity.Flat"
                    }
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Login": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_controller_v1_response.Registration": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_romanchechyotkin_avito_test_task_internal_entity.Flat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "houseID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderationStatus": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "roomsAmount": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}